Create powerline_SQL_script
Because of the variable nature of the data, most of the effort goes into standardizing column data and splitting it into more digestible pieces that can be compared. REGEXP are very handy here as they can handle lots of conditions and are very flexible in identifying patterns. 

The script is also written in a way that additional cases can easily be added or removed if desired for future data sets. This is very handy for making it a dynamic solution to a reoccurring problem.